Add support for config.maxlength so records that are too large are truncated
This isn't great, it applies the maxlength to each "selector" in the config file and Algolia limits the total size of a record to 10,000 bytes. A value of 8500 seems to work for our purposes. Also since each record[key] is an array of all the elements that match the selector this pops the last element off oversize selectors until it is under the limit. If there's one really big element you will end up with an empty selector. It would be better at the end of the parse function and evaluating the size of the entire record.

Also adds URLs to error responses that didn't have them.